manque_pas_grand

+++ ft_printf.h
@@ -15,7 +15,7 @@ var_list var;
 int ft_printf(const char *format, ...);
 void check_format_sp(char c, va_list args);
 void ft_putnbr_unsigned_fd(unsigned int u, int fd);
-void ft_itoa_to_hexa(long unsigned int nb, char c);
+void ft_itoa_to_hexa(long long nb, char c);
 
 
 #endif
